added eureka and api-gateway for user-service, room-service and partly for hotel-service (as not all sub-services are implemented), reformatted pom's, changed cost variable in room from int to double

+++ room-service/src/main/java/soa/roomservice/dto/RoomRequest.java
@@ -15,6 +15,6 @@ public class RoomRequest {
     private String roomNumber;
     private String description;
     private Integer capacity;
-    private int cost;
+    private Double cost;
     private RoomType type;
 }
